NIFI-10159 Move internal interfaces from c2-client-api
