+++ minifi/minifi-bootstrap/src/test/java/org/apache/nifi/minifi/bootstrap/configuration/differentiators/WholeConfigDifferentiatorTest.java
@@ -31,8 +31,7 @@ import java.util.Properties;
 import java.util.concurrent.atomic.AtomicReference;
 import okhttp3.Request;
 import org.apache.commons.io.FileUtils;
-import org.apache.nifi.c2.client.api.ConfigurationFileHolder;
-import org.apache.nifi.c2.client.api.Differentiator;
+import org.apache.nifi.minifi.bootstrap.ConfigurationFileHolder;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
